Moved a bunch of classes into a folder
