+++ LitDev/LitDev/Finances/Company.cs
@@ -1,0 +1,8 @@
+﻿namespace LitDev.Finances
+{
+    public class Company
+    {
+        public string symbol;
+        public Profile profile;
+    }
+}
