+++ LitDev/LitDev/Finances/FinancialWrapper.cs
@@ -1,0 +1,8 @@
+﻿namespace LitDev.Finances
+{
+    public class FinancialWrapper<T>
+    {
+        public string symbol;
+        public T[] financials; 
+    }
+}
