+++ LitDev/LitDev/Finances/Price.cs
@@ -1,0 +1,8 @@
+﻿namespace LitDev.Finances
+{
+    public class Price
+    {
+        public string symbol;
+        public decimal price;
+    }
+}
