+++ LitDev/LitDev/Finances/Profile.cs
@@ -1,0 +1,24 @@
+﻿namespace LitDev.Finances
+{
+    /// <summary>
+    /// Information about a companies profile
+    /// </summary>
+    public class Profile
+    {
+        public decimal price;
+        public decimal beta;
+        public int volAvg;
+        public double mktCap;
+        public decimal lastDiv;
+        public string range;
+        public decimal changes;
+        public string changesPercentage;
+        public string companyName;
+        public string exchange;
+        public string industry;
+        public string description;
+        public string ceo;
+        public string sector;
+        public string image;
+    }
+}
