+++ LitDev/LitDev/Finances/Search.cs
@@ -1,0 +1,14 @@
+﻿using Newtonsoft.Json;
+
+namespace LitDev.Finances
+{
+    public class Search
+    {
+        public string symbol;
+        public string name;
+        public string currency;
+        public string stockExchange;
+        [JsonProperty("exchangeShortName")]
+        public string stockExchangeShortName; 
+    }
+}
