+++ LitDevUnitTests/Finances.cs
@@ -5,6 +5,7 @@ using System.Text;
 using System.Threading.Tasks;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using NUnit.Framework;
+using LitDev.Finances;
 using Assert = Microsoft.VisualStudio.TestTools.UnitTesting.Assert;
 
 namespace LitDevUnitTests
@@ -15,7 +16,7 @@ namespace LitDevUnitTests
         [TestMethod]
         public void GetProfile()
         {
-            var data = LitDev.Finances.GetCompanyProfile("AAPL");
+            var data = Engine.GetCompanyProfile("AAPL");
             Assert.AreEqual("AAPL", data.symbol);
             Assert.AreEqual("Technology",data.profile.sector);
         }
@@ -23,7 +24,7 @@ namespace LitDevUnitTests
         [TestMethod]
         public void GetQuote()
         {
-            var data = LitDev.Finances.GetCompanyQuote("AAPL");
+            var data = Engine.GetCompanyQuote("AAPL");
             Assert.AreEqual("AAPL", data[0].symbol);
             Assert.AreEqual("Apple Inc.", data[0].name);
             Assert.AreEqual("NASDAQ", data[0].exchange);
@@ -32,7 +33,7 @@ namespace LitDevUnitTests
         [TestMethod]
         public void GetSearch()
         {
-            var data = LitDev.Finances.GetSearch("AA", "NASDAQ");
+            var data = Engine.GetSearch("AA", "NASDAQ");
             Assert.IsNotNull(data[0]);
             Assert.IsNotNull(data[9]);
         }
@@ -40,12 +41,12 @@ namespace LitDevUnitTests
         [TestMethod]
         public void GetFinancialStatements()
         {
-            var annual = LitDev.Finances.GetIncomeStatement("AAPL");
+            var annual = Engine.GetIncomeStatement("AAPL");
             Assert.AreEqual("AAPL", annual.symbol);
             Assert.IsNotNull(annual.financials);
             Assert.IsTrue( annual.financials[0].date.CompareTo( new DateTime(2019, 09, 8)) >= 0 );
 
-            var quarterly = LitDev.Finances.GetIncomeStatement("AAPL", LitDev.Finances.ReportingPeriod.Quarterly);
+            var quarterly = Engine.GetIncomeStatement("AAPL", Engine.ReportingPeriod.Quarterly);
             Assert.AreEqual("AAPL", quarterly.symbol);
             Assert.IsNotNull(quarterly.financials);
             Assert.IsTrue(quarterly.financials[0].date.CompareTo(new DateTime(2020, 03, 28)) >= 0);
@@ -54,7 +55,7 @@ namespace LitDevUnitTests
         [TestMethod]
         public void GetBalanceStatements()
         {
-            var annual = LitDev.Finances.GetBalanceStatement("AAPL");
+            var annual = Engine.GetBalanceStatement("AAPL");
             Assert.AreEqual("AAPL", annual.symbol);
             Assert.IsNotNull(annual.financials);
         }
@@ -62,7 +63,7 @@ namespace LitDevUnitTests
         [TestMethod]
         public void GetCashFlowStatement()
         {
-            var annual = LitDev.Finances.GetCashFlowStatement("AAPL");
+            var annual = Engine.GetCashFlowStatement("AAPL");
             Assert.AreEqual("AAPL", annual.symbol);
             Assert.IsNotNull(annual.financials);
         }
@@ -70,7 +71,7 @@ namespace LitDevUnitTests
         [TestMethod]
         public void GetStockPrice()
         {
-            var price = LitDev.Finances.GetRealTimePrice("AAPL");
+            var price = Engine.GetRealTimePrice("AAPL");
             Assert.AreEqual("AAPL", price.symbol);
             Assert.IsTrue(price.price > 0);
         }
